fix: rally segmentation with shot outcome detection + pipeline regression fixes
Rally segmentation:
- Add _is_rally_ending_shot() to detect rally endings via shot outcomes
- Uses stroke type + confidence + gap as signals (not just time gap)
- High-confidence smash/drop/kill with gap > 25 frames ends rally
- Net shots with gap > 15 frames end rally
- Large gap (> 45 frames) always ends rally

Pipeline regression fixes:
- Hit threshold 85th -> 90th percentile (reduces false hits)
- Rally gap 45 -> 60 frames (better rally separation)
- BST softmax computed in FP32 (fixes zero-confidence issue)
- Speed peaks distance 5 -> 3 (matches v1.1)
- Dedup window 5 -> 8 frames (better separation)

Tests:
- Add 5 new tests for _is_rally_ending_shot()
- Update existing rally tests for new behavior

+++ backend/app/models/bst.py
@@ -143,7 +143,7 @@ class BSTClassifier:
 
                 with torch.no_grad():
                     logits = self.model(JnB, shuttle, pos, video_len)
-                    probs = torch.softmax(logits, dim=1).cpu().numpy()
+                    probs = torch.softmax(logits.float(), dim=1).cpu().numpy()
 
                 for j in range(len(batch_clips)):
                     pred_idx = int(np.argmax(probs[j]))
